socrates_llm/pyproject.toml
[project]
name = "socrates-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.2.0",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "torch>=2.5.1",
    "tqdm>=4.67.1",
    "transformers>=4.47.1",
    "trl>=0.13.0",
]